Bugfix: don't wipe event signups unless relevant changes were made

+++ api/src/main/java/net/blueshell/api/listener/jpa/QuestionEventListener.java
@@ -15,6 +15,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
+import java.util.Set;
+
 @Slf4j
 @Component
 @RequiredArgsConstructor
@@ -23,17 +25,25 @@ public class QuestionEventListener {
     private final AnswerService answers;
     private final EventSignUpService signUps;
 
+    private boolean relevantChanges(Set<String> changes) {
+        return changes.contains("choiceLabels") || changes.contains("label") || changes.contains("type");
+    }
+
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void onPersist(PostPersistEvent<Question> evt) {
         var q = evt.getSource();
+        var changes = evt.getChangedProperties();
 
         log.info("On persist question {}", q);
 
         // If a new description is added, there is no need to clear the survey.
         // If a new question is added, then we do need to wipe the answers and signups.
         // This is because the surveys will need to be filled in again.
-        if (q.getType() != QuestionType.DESCRIPTION && q.getSurveyId() != null) {
+        if (q.getType() != QuestionType.DESCRIPTION
+                && q.getSurveyId() != null
+                && relevantChanges(changes)
+        ) {
             signUps.deleteAll(signUps.findBySurveyId(q.getSurveyId()));
         }
     }
@@ -42,12 +52,17 @@ public class QuestionEventListener {
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void onUpdate(PostUpdateEvent<Question> evt) {
         var q = evt.getSource();
+        var changes = evt.getChangedProperties();
 
         log.info("On update question {}", q);
+        log.info("On update changed properties {}", changes);
 
         // When a question is updated, the survey will need to be re-filled.
         // Therefore, all answers for the survey need to be wiped.
-        if (q.getAnswers() != null && !q.getAnswers().isEmpty()) {
+        if (q.getType() != QuestionType.DESCRIPTION
+                && q.getSurveyId() != null
+                && relevantChanges(changes)
+        ) {
             signUps.deleteAll(signUps.findBySurveyId(q.getSurveyId()));
         }
     }
